main-app/backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9

# WebSocket
websockets>=12.0
python-socketio>=5.11.1

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
aiosqlite>=0.19.0
greenlet>=3.0.0

# Authentication
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
ldap3>=2.9.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0

# LLM Integration (versions managed by langchain-openai / langchain-anthropic)
openai>=1.12.0
anthropic>=0.18.1
tiktoken>=0.6.0

# Agent Framework (LangGraph CodeAct)
langchain-core>=0.3.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.4.0

# Agent Tools
ddgs>=9.10.0
Pillow>=10.0
pyautogui>=0.9
trafilatura>=1.12.0
tavily-python>=0.5.0

# Browser Automation (Playwright for AutoBrowser tool)
playwright>=1.49.0

# Document Processing (PyMuPDF is faster for PDFs; pdfplumber is fallback)
pymupdf>=1.24.0
pdfplumber>=0.10.4
python-docx>=1.1.0
mistune>=3.0.2
beautifulsoup4>=4.12.3

# OCR (optional — graceful degradation if not installed)
pytesseract>=0.3.10
pdf2image>=1.16.3

# Numerical
numpy>=1.24.0,<2.0.0

# Excel support (for ticket export processing)
openpyxl>=3.1.2

# Email Provider OAuth
msal>=1.28.0
google-auth>=2.29.0
cryptography>=42.0.0

# Utilities
httpx>=0.27.0
structlog>=24.1.0
python-dotenv>=1.0.1
orjson>=3.9.13
aiofiles>=23.2.1
PyYAML>=6.0.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0

# Rate Limiting
slowapi>=0.1.9

# MFA
pyotp>=2.9.0
qrcode[pil]>=7.4.2

# Development
black>=24.1.1
ruff>=0.2.1
mypy>=1.8.0
apscheduler>=3.11.0
pytz>=2024.1

# MCP Builder
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
